Fixed style-loader
Style-loader is recommended for dev over MiniCssExtractPlugin as it is faster. However it is incompatible with server builds and had seemingly been partially removed in Razvan's changes. This commit re-introduces it bringing it more in line with the original webpack files before the attempted Bower removal.

Due to this base config is now generated on the fly using a flag that determines whether the build is a client or server build.

For server builds, both style-loader (inline style tags added to the HTML) and MiniCssExtractPlugin (generates css files and adds them to the DOM) are omitted.

For client builds, one is present depending on whether it is a prod build or a dev build.