Tc 2503/ssr recipecards as entry point (#3591)
* [TC-2503] Get rid of routes.js for Ingredients

* [TC-2503] Include orderId and userId in the getHelp recipeCard route

* [TC-2503] Stop using content coming from the store

Now it's hardcoded. We were using harcoded defaults anyway because the development of our Content Service halted.

* [TC-2503] Remove the presentation component

This component does not have logic at all, and in those case we don't have one logic component and one presentation component, but only a logic one.

* [TC-2503] Update the layout to GetHelpLayout2

* [TC-2503] Change the style of Ingredients, in the previous ticket I just made it bold and that doesn't truly matched what there was before

* [TC-2503] Improve the tests

- Remove warnings
- Change reset mocks to clear mocks. It is more lightweight, so we should use reset only if clear is not enough for our case.
- As we are clearing the mocks (same if we were resetting) we don't need to create new mock functions before each test
- Add a missing test
- Stop naming mocks as 'mocks', as we decided to stop using that convention
- Change recipe ingredients ids to uuids to match proptypes

* [TC-2503] Add cypress test for direct URL access to GetHelp Recipe Cards

* [TC-2503]/CR Fix a test description

* [TC-2503] Use selector to get data from the store in loadOrderAndRecipesByIds action

Also, I updated all the places that could use the selector. And there was a selector already for the recipes, in the general selectors folder, but we want to move everything to our domain in the codebase, so I deleted that selector and used the new one where needed