[TG-4320] Apply promo code on the menu page (#3601)
* [TG-4320] remove checking props and location from getPromoBannerState

Location checking never worked since 2f8f6494b5c7fcd577089b74ed94c8f5fafcaa1d
(that commit moved from checking window.location to checking ownProps.location,
but didn't introduce passing location to PromoBanner props).

Since nowadays when you put promo code in the query parameter, the query
handler works early and saves the code to the store, this check isn't needed
anyway.

Likewise for ownProps: PromoBanner never received a promoCode prop (since at
least 2019), so checking it does nothing.

* [TG-4320] remove promoApplicable utility

Since `applyIf` was never anything but 'loggedOut', it doesn't make sense to
have code dealing with possible other three states.

* [TG-4320] rename mayCode to defaultPromoCode

`may` is not clear enough to describe the purpose of the value.

* [TG-4320] remove check for non-null promoCode

Since promoCode is taken either from feature flag, or from config
defaultPromoCode, it can never be empty.  Hence the check is unnecessary.

* [TG-4320] remove hide and leave only canApplyPromo to signify !hide

Since

a) in most places we used `!hide && canApplyPromo` to detect the fact that
"promo banner is visible and promo is applicable" in order to apply the default
promo code on navigation; and

b) after the simplifications it turned out that this condition is equivalent to
just `!hide` (since `canApplyPromo` was depending only on `isAuthenticated`,
which was one of the legs for `hide`)

It didn't make sense to keep passing both `hide` and `canApplyPromo` around.

This commit redefines the meaning of "canApplyPromo" to be the old `!hide &&
canApplyPromo` condition and to mean the same as "!hide".  Hence `hide` became
redundant; we now only return "canApplyPromo"; in all places where we checked
"hide" we now use "!canApplyPromo" instead.

* [TG-4320] rename applyPromoCodeAndRedirect -> applyPromoCodeAndShowModal

in preparation of removing the redirect functionality out of it

* [TG-4320] remove redirect functionality out of applyPromoCodeAndShowModal

Instead the calling places do a redirect whether this helper was successful or not.

* [TG-4320] applyPromoCodeAndShowModal is now an async action instead of helper

Because it is exported as an async action and used in one place as such, in
order to maintain the principle of least surprise it is now an action. (And
the other place where it was used was changed to use the action calling style.)

* [TG-4320] apply default promo code and show modal on the menu page

* [TG-4320] Update regression tests

* [TG-4320] Remove commented code

* [TG-4320] update e2e tests: dismiss promo after landing on menu

* [TG-4320] code health

* [TG-4320] fixed incorrect unit test

* [TG-4320] reuse the action instead of repeating its logic

* [TG-4320] rename mayText to defaultBannerText; remove unused value

Co-authored-by: zinkevichav <alexandra.zinkevich@gousto.co.uk>