Tc 2489/redesign next order (#3665)
* [TC-2489] Pass the whole order instead of orderId to NextOrder component

Additionally, instead of using a generic proptype, a detailed proptype for order has been created.

However, I made a mistake and use the wrong order (a different place in the store). As the work was done, I used the wrong order props in a place that uses that order, so it does not go to waste.
In the next commit I amend this mistake.

* [TC-2489] NextOrder redesign

Some notes:
- I made a mistake with the order proptypes (different order in the store). However as the work was done, I used them somewhere rather than remove them. In this commit I add the right ones and use them in NextOrder and Header instead of the wrong ones from the previous commit
- The primary and secondary message props from Next Order have been removed, but further clean up in other components will be done in other commit.

* [TC-2489] Use order proptype from a single source of truth

* [TC-2489] Get rid of linkUrl and linkLabel props and calculate them inside NextOrder

linkUrl and linkLabel are now hardcoded props in NoNextOrder, and they will be removed and done inside (and updated) in the ticket that tackles the redesign of that component.

* [TC-2489] Get rid of primaryMessage and secondaryMessage in Header

They had been removed from NextOrder in a previous commit, so they are not needed.
The `hasNextOrder` bool was unexplicably depending on secondaryMessage being present. Instead, now we are using `nextOrder` to determine if it's there or not

* [TC-2489] Fix Cypress testing selector

* [TC-2489]CR avoid using a no-op function

* [TC-2489]CR Move some tests out of the __tests__ folder

We prefer to completely colocate them now

* [TC-2489]CR Add and use a util to obtain the order state we display in the clients

* [TC-2489]CR make the non tracking button primary in some circumstances

* [TC-2489]CR Move the tracking actions from Header container to NextOrder container