Rewrote the entire pipeline
# Old Pipeline
`setup-source`, included steps to do an install global modules an npm ci and an `npm run postinstall` which installed the bower modules. It did this in preparation for the code health steps. This was actually all redundant as all of the subsequent steps require webclient to be built and deployed and these steps perform all the same actions again.

`setup-platform` used to execute `platform-setup.sh` from the gousto-platform-deploy repo to download all the other platform scripts and set the `ENVIRONMENT` environment variable from the branch name.

`setup-build-application` used to do various environment variable set up (`setup-common.sh`) and `build-application` ran `setup-isomorphic.sh` which built webclient e2e helpers, then `build-webclient-image` would upload the built webclient files to S3.

The `deploy-application` used to execute `deploy.sh` which I think was a hangover from G2FE. It did an npm upload and then an npm ci (again) for some reason in the dist directory after copying in node_modules, it then downloaded the bower modules again and then ran `deploy-service.py` which created half of the infrastructure (this is the modern gousto-platform way of doing it).

The `deploy-webclient-image` step used to create the webclient docker image post build and upload it to the ECR. This has been renamed for clarity.

The `deploy-containers` step used to run the platform `deploy-containers.py` script that would deploy the last half of the architecture and then deploy the previously uploaded docker image from the ECR to ECS via Fargate.

# New Pipeline
These steps are listed by and large in chronological order with five main steps.

## Setup
The `setup-source` and `setup-platform` steps were merged in to single setup step (` setup-ci-variables-and-download-platform-scripts`) for ci variables such as `ENVIRONMENT` which is set by a platform script (`platform-setup.sh`) from the branch name and controls which ECS cluster etc it gets deployed to and our various build steps. It also downloads platform scripts and now sets up the Elastic Container Registry (ECR) in advance, where the webclient image is uploaded for deployment, which used to happen in the `setup-platform` step.

## Build
A lot of the deploy and build steps have been now merged in the new `build-webclient`, it builds the webclient from source after doing an npm install and then uploads it to s3, that's it.

## Unit test in parallel with deploy
After `build-webclient`, `eslint-generate-report` and `jest-coverage-report` are ran in parallel with all the below deployment steps. Since we already have ran this prior to merge of the branch, they should not fail. The `jest-coverage-report` step runs the CI tests along with a coverage report. The results of this are then published in `public-code-health` so that they can be used in the the Github Actions pipeline per branch.

## Deploy
Then `create-and-upload-webclient-docker-image-to-ecr` does as its name suggests and uploads it to ECR to get ready for the deployment to ECS.

The deployment to ECS along with other infrastructure deployment steps happens after in the `deploy-infrastructure-and-webclient-image` step after all the code and images are built and uploaded. This is where the actual webclient docker image is deployed to ECS and started.

## Regression and E2E
Finally there is the `install-regression-test-dependencies` that does an npm ci for regression tests before any of them execute and `test-e2e-selenium` now builds the webclient e2e helpers itself and the e2e tests themselves before leasing IPs from the WAF so that the CI node can access webclient for account setup etc. It then runs the tests. At the end, `test-e2e-selenium`, `test-regression` and `test-regression-mobile` all run in parallel similar to how they used to.

# Improvements
* Remove `deploy-containers.py` or `deploy-service.py` steps whichever is easier. Really should be using deploy-service which is the one maintained by Potatoes. This would yield a 4 minute performance improvement and I'm not convinced they can be easily parallelised as running them in the opposite order and then reversing them seemed to need me to tear down the CFn stack. If ran in parallel there would be no real control of which finished first. In the current pipeline they are technically ran in parallel but because so much happens before `deploy-containers`, `deploy-platform` (`deploy-service`) consistently finishes first.
* Use either pre-built CircleCI docker build images where possible or upload the Gousto build images to `us-east-1` AWS region. This will massively speed up the 30+ seconds per workflow step taken to download the docker image that executes those steps
* Use a [ramdisk](https://support.circleci.com/hc/en-us/articles/360054908812-Speed-up-steps-using-a-RAM-disk)
* Remove the S3 upload as Cloudfront should adequately cache assets that the server can serve initially. Load might increase initially on the Express server. But this would only save about 6s on the build time, it just makes a little more logical sense.
* Build a multi stage docker image and reuse it throughout the build process to run steps, this would prevent persisting the circle CI workspace between each workflow step and also be an image to deploy straight away rather than building and then making a docker image with the built files.
* Remove redundant webpack modules and work out why SASS needs rebuilding all the time. The architecture for building can't be that unusual.

With these additions sub 15 minutes for a staging deployment might be possible.