Tc 2618/get help next order projected deliveries (#3711)
* [TC-2618] Add loadNextProjectedOrder action

* [TC-2618] Add myGousto reducer

* [TC-2618] Add getNextProjectedOrder selector

* [TC-2618] Create NextProjectedDelivery component

* [TC-2618] Render next projected delivery when available

* [TC-2618] Make the projected deliveries load correctly when hard reloading My Gousto

* [TC-2618] Fix test warning

* [TC-2618] Add action to get orders for My Gousto

Initially, we were using the global one. Instead we are creating one for My Gousto domain.
The problem is that the global corresponding reducer merges the orders obtained from the BE with the previous orders that obtained before, effectively providing out of date orders.

* [TC-2618] Add reducer for actiontype MY_GOUSTO_LOAD_ORDERS

Initially, we were using the global userLoadOrders and its reducer. Instead we are creating one for My Gousto domain.
The problem is that the global corresponding reducer merges the orders obtained from the BE with the previous orders that obtained before, effectively providing out of date orders

* [TC-2618] Add selectors for previousOrder and nextOrder in My Gousto

* [TC-2618] Add a Loading spinner and fix out of date orders being shown.

In order to fix out of date orders being shown, we used our own set of (action, reducer, selectors) to get the orders.
We fetch them when Header loads, and render the Loading spinner until we get the orders and projected deliveries.

I took the opportunity to simplify the Header component, instead of passing a bunch of orders and letting it find the previous and next orders, that's now the reducer responsibility and Header receives next and previous order directly.

* [TC-2681]CR Improvements to loadNextProjectedOrder action

* [TC-2618]CR Refactor to avoid nested ternary statements

* [TC-2618]CR Fix typo in test description

* [TC-2618]CR Fix eslint error

* [TC-2618]CR Raise loadNextProjectedOrder action coverage to 100%