Boosted docker build container resources
Probably not required for all steps, but put on to all steps to see the benefits.

Using the ramdisk seemed to cause some steps to slow down or hang such as the jest tests, probably due to lack of available resources. This will hopefully alleviate this.

However this commit won't actually become useful until we upgrade our CircleCI plan to allow us to use the larger executors with more resources.