Fix path sanitisation of code files when collating coverage and test reports (#3901)
Previous code was heavily tied to the folder structure, this removed /gousto-webclient and /root/project from the code paths when merging the Jest and ESLint reports before publishing the results for branch comparison. This broke with the pipeline change to move to running these reporting steps in /mnt/ramdisk rather than /root/project.

This change still is somewhat tied to the folder structure and this is not great, it's a quick fix. However at least it doesn't try to remove parts of the path, it just returns the part of the path that follows `src/`. This of course assumes our code is in a `src/` directory, but that is a more fair assumption than assuming the path will always contain /root/project.