[FEF-172-HF] - Fix code path sanitisation for code health again (#3909)
Sadly when the branch GHA pipeline is running, it uses the function with an entire list of files that have changed, many of which aren't in the src directory. This results in match returning null and a TypeError accessing [0] of null. This method rewrite reinstates the original behaviour present in the previous code using `String.replace` which would return the original string if nothing could be stripped.