Typed fetch.ts
Did my best here as the fetch code deals with some helpers that return really inconsistent payloads and errors.

This required more permission union types rather than intersection ones and some awkward type casting (in the case of Lumberjack errors). Either that or lots of guard clauses.

This has been left to show how TS can highlight messy code and how it really needs to be refactored.

The timeout functionality has been removed as it wasn't used and only works on the server with node-fetch. In fact it couldn't have even been working before as the version of node-fetch used by our version of isomorphic-fetch prior to the upgrade didn't have this feature.

Added a globals file.