Added typescript-eslint
This adds the typescript-eslint parser which is compatible with JS. Also updated ESLint.

Ignoring TS files and Typescript specific config produces similar results and warnings as the old babel-parser.

In fact, it is a little more strict and picked up a few more things, such as unnecessary curly braces.

Note the no-use-before-define has been disabled in favour of the TS one due to compatibility reasons as per: https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-use-before-define.md

Comes with a sane config for linting TS files only. Ideally this should be merged in to the JS config, but the JS one is a complete mess at the moment.

Another thing to note is the parserOptions which for TS files takes into account the tsconfig.json but does not for JS files. Including tsconfig.json this enables tighter checks that require properties from that file, but for some reason when parsing js files ESLint hangs. The only relevant issue I could find was this: https://github.com/typescript-eslint/typescript-eslint/issues/1500 and the suggestions failed to resolve the issue.

On a further look, maybe the solution can be found here: https://github.com/eslint/eslint/issues/13848