[TG-5194] re-enable webpack-bundle-analyzer (#4122)
* [TG-5194] delete trailing whitespace

* [TG-5194] enable webpack-bundle-analyzer on prod builds

* [TG-5194] re-add run instructions to readme

* [TG-5194] collect stats only if requested via an environment variable

The bundle size stats are most correct when they're collected against a
production bundle.  On the other hand, we don't want to run it in CI because
that increases build time and upload the useless >100MB file to S3.  Hence the
plugin is enabled only when building locally for the purpose of collecting
the stats.