[TG-5224] Remove some warnings in build (#4136)
* [TG-5224] remove unused quotes file

Its only usages were removed in 2a0f33283fb5652fb83274d6ba97aebcc429bebf, but
nobody bothered to delete the file itself.

* [TG-5224] append "live" to config file names to match NODE_APP_INSTANCE value

This gets rid of the warning

```
WARNING: NODE_APP_INSTANCE value of 'live' did not match any instance config file names.
WARNING: See https://github.com/lorenwest/node-config/wiki/Strict-Mode
```

* [TG-5224] remove collecting `GIT_HASH`

This gets rid of the "not a git repo" error in the build logs.

It was introduced only for the purposes of including source maps for
Raven.js (for Sentry) in gousto2-frontend
50a65fdc889f03d1a63a8d3a4ce361a278c278d0.

Raven.js was removed in 815470931a35ddc9fa853b2376159bceae48ae47; so this piece
of data isn't used anywhere.

* [TG-5224] move browserlist config to .browserlistrc

This gets rid of the warning about its use.

* [TG-5224] initialize plugins only when requested to avoid side effects

In production build, there were two lines `Webpack: Starting ...` that were
output due to inclusion of SimpleProgressWebpackPlugin, even though this plugin
is meant to be used in the development mode only.  So plugin initialization is
now in a function that is invoked on demand (rather than in the top-level
directly, where both development and production plugins were initialized at the
same time in both development and production builds).

* Revert "[TG-5224] append "live" to config file names to match NODE_APP_INSTANCE value"

This reverts commit dd3542443189929a42eb8e79a37461c0d2e9a145.

As it breaks the expectations about how `node-config` should be used, this
change was poor.  Better to live with this warning for a while more - and the
whole deal with the huge repeated configs will be cleaned up with the move to
BODA that's in FEF's plans.