[TG-4731] upgrade redux to v7 (#3796)
* Update redux version

* update classes to use contextType pointing to ReactReduxContext

* fixup! update classes to use contextType pointing to ReactReduxContext

* [update-redux] fix Link navigation: return use of react-router context

Converting to use ReactReduxContext was a mistake, because Link uses the
react-router context and doesn't care about redux.

* [update-redux] fix more proptype warnings

* [TG-4731] replace withRef with forwardRef as per redux-form v8 docs

https://github.com/redux-form/redux-form/releases/tag/v8.0.0

(Not that it helps with anything from what I see.)

* [update-redux] remove CheckoutPaymentContainer aliasing

* [TG-4731] CheckoutPaymentContainer set up like other form sections

Calling formContainer() on result of connect() caused react renderer to
complain that "type" is undefined on a memo component.  Instead we call
formContainer() on a component which renders the same result of connect().  I'm
not sure why this works, and the form setup code is still ridiculously complex,
but it seems effective for now.

* [TG-4731] fix codehealth warnings after rebase

Co-authored-by: Jason Lafferty <jason.lafferty@gousto.co.uk>
Co-authored-by: Alexander Rymasheusky <alexander.rymashevskiy@gousto.co.uk>