[TG-5512] fix "Stale element reference" error in e2e tests
b967cb731cc65279f0a12db5d5f1cfd93be4a1a3 introduced a regression: before,
the Add Recipe buttons selector was run before clicking each button; after, it
was run once, then resulting ids were iterated for clicking.  With the recipe
cards rerendering after each addition of recipe into the basket, the selenium
view of the element ids has the chance to change, resulting in the element
references becoming stale.  This commit reinstates the previous behavior of
getting the up-to-date element references (in addition to the useful bits of the
offending commit, namely dismissing all kinds of follow-on pop-ups that can
appear as a result of adding the recipe).