Prelim for FEF-173: begin typing JSONHelper (#4262)
This pull request takes on typing the JSONHelper util, which should help with FEF-173 (typing the fetch util)

### What's JSONHelper, and why do we want to type it?

JSONHelper is used pretty much everywhere we do an API call, so providing types for its functions will be of benefit across the app.

Many server responses are passed through `parseObjectKeysToCamelCase`, which is now typed. We use the new string template types from TS4.2 to represent a snake_case -> camelCase conversion.

Of course, this only helps if the _input_ to the function is typed, but if you can do that, you now get types like the following:

<img width="672" alt="image" src="https://user-images.githubusercontent.com/3148617/140976244-5ce5b423-1408-439b-8f5d-abaebc14ef40.png">

### So we get to completely rewrite it in TypeScript, right?

Not exactly. The source file is 1) moderately complex from a code perspective; 2) quite complex from a types perspective; 3) absolutely critical to most functionality in the app. For that reason, I've trod carefully and kept the implementation and semantics almost the same, except in trivial cases.

That has meant I've opted to use heavier type assertions than I'd normally like, if it lets me keep a one-to-one code mapping. The `parseObjectKeysToCamelCase` is almost unchanged except for types. Only `formatErrorsWithCode` is really 'refactored'.

### I notice `processJSON` is still in JS. Did you forget about that?

This PR does *not* port the processJSON function - Ash (@Astridax) and I looked at this before and it was quite complex, requiring a decent representation of the different API response types available. I think we should do that next, in a separate PR, and take this work in small chunks.

### Some of the types look very complicated. How do we test them?

The types for mapping snake case to camel case are a little complicated, so this PR adds a utility for testing complex mapped types. This will stop the tests compiling if a mapped type isn't the shape we expect.

Here's how it looks in CI if a type test fails:

<img width="920" alt="image" src="https://user-images.githubusercontent.com/3148617/140977102-6e95f6e7-85f8-4912-b99b-f9e1fa3f4cce.png">

### How do we move to TypeScript long term?

In my opinion: small steps, high confidence. Better to try and "fix" one file once than two files twice. That gives us breathing room to discuss problems we face and not overwhelm ourselves. Or break the website...

Whilst I get the excitement to start using TS ASAP, with a team who's still learning TS, I think it's best to start steady and figure out good patterns. Then we can all know that the example TS we're looking at has been agreed on and isn't likely to change.