[TG-5673] Remove unnecessary requests on the Home page (#4452)
* [TG-5673] remove code related to goToMyGousto/goToMyDeliveries

Introduced in f7c17f0ab0c0bfead4f83acd51d3b83ede8a26dc - no links, no
documentation, no context.

The a/b test was running in spring 2019, see

https://app.optimizely.com/v2/projects/3732210624/experiments/14109380660/variations/14129770123

It was archived and forgotten, but the code handling it kept on going.

* [TG-5673] fix AppBanner prop type warnings

name and ratings cannot be required props for the components because it works
fine without them when showAppBanner is false.

* [TG-5673] move ordering logic to the selector

It makes no sense to invoke selector, then call the extra processing logic in
the presentational component with the result of the selector.  Why not return
what the component needs in the first place?

* [TG-5673] remove unnecessary container and alias

* [TG-5673] retrieve from the menu only what's needed for carousel

Previously the whole menuFetchData flow was requested, which makes separate
calls to the menus, to the brand data, to the delivery days; all of this is
wasted when our only purpose is to get recipes and collections and put
non-clickable recipe preview cards into a carousel.

Additionally, ingredients are not requested from the menu service on the
assumption that the endpoint will do less work if it doesn't need to retrieve
the ingredients.  (And the client will definitely do less work not having to
parse the ingredients JSON and stuff it into the store.)

* [TG-5673] display recipe placeholders of about the same height for less CLS

* [TG-5673] don't show cursor: pointer because recipes are not clickable

* [TG-5673] use requestIdleCallback to delay the loading of carousel recipes

* [TG-5673] skip fetching subscription pause data for prospects

Since the prospects don't have a subscription, we cannot have a reason for
pausing it, so there's no need to make the requests at all.  This gets rid of
two requests ending with "401 Unauthorized" made at any pages that include the
full Header.

* [TG-5673] fix proptype warning for className prop om BoxSummary/Postcode

* [TG-5673] fix proptype warning for react-modal/Modal className

* [TG-5673] update Home unit tests

* [TG-5673] update SubscriptionPause unit tests

* [TG-5673] fix some eslint warnings

* [TG-5673] add a comment for the requestIdleCallback check