[TG-6040] update eslint, start checking TS files (#4876)
* [TG-6040] upgrade eslint and its plugins (and start checking ts files)

This gets rid of the following error:

```
% npx eslint server/datadog.ts
Error while loading rule '@typescript-eslint/no-loss-of-precision': @typescript-eslint/no-loss-of-precision requires at least ESLint v7.1.0
Occurred while linting /Users/rymashevskiyas/home/projects/gousto-webclient/src/apps/webclient/server/datadog.ts
```

Incidentally, and for no reason I can explain, it starts checking the ts files
under src.

* [TG-6040] fix checking ts files in server and src/.../__tests__

This gets rid of the following errors:

```
npx eslint server/__tests__/datadog.test.ts
  0:0  error  Parsing error: "parserOptions.project" has been set for @typescript-eslint/parser.
The file does not match your project config: server/__tests__/datadog.test.ts.
The file must be included in at least one of the projects provided
```

```
npx eslint src/utils/types/__tests__/array.test.ts
  0:0  error  Parsing error: "parserOptions.project" has been set for @typescript-eslint/parser.
The file does not match your project config: src/utils/types/__tests__/array.test.ts.
The file must be included in at least one of the projects provided
```

* [TG-6040] disable no-promise-executor-return check

The pattern of
```
new Promise((resolve) => resolve({ data: { isRecaptchaEnabled: false }}))
```

is perfectly fine.

* [TG-6040] suppress default-param-last error

This is indeed a code error but I cannot fix it now.

* [TG-6040] update configs to guard changes in ts/tsx files in Menu

This will only touch ts/tsx files because doing all of routes/Menu in one go
changes 360 files which is not for the faint of heart; JS migration is
a separate step - TG-6078.

* [TG-6040] don't double-report `"` issues

* [TG-6040] fix compare-code-health output

eslint 8 stopped exporting the formatters directly; but it can still be
accessed via the API; and `stylish` is the default choice.

* [TG-6040] suppress some warnings for less noise

* [TG-6040] configure camelcase to allow the established patterns

* [TG-6040] change configs to warn of some issues and ignore others

The actual issues that are present are downgraded from errors to warnings to be
dealt with later.  The formatting issues are ignored on ts files because
Prettier will catch them soon.  no-underscore-dangle is allowed becaus is
schizophrenic to allow undefined variables with the underscore prefix but
forbid underscore dangling.

* [TG-6040] fix and suppress some issues in ts files

This overlaps a bit with the work to be done in TG-6082, but I'd like the
current PR to output no eslint "errors" so that the pipeline doesn't fail - so
the stuff that could be fixed/suppressed easily was done in this PR.  (And what
couldn't was marked as warnings in the previous commit.)

* [TG-6040] fix some auto-fixable issues in js files

I have no idea why they didn't get reported before.

* [TG-6040] fix some indent errors