[TG-6317] remove <form> tag so Safari AutoFill cannot submit it incorrectly (#5034)
I cannot find official reference, but from the sources I could [1] apparently
the AutoFill feature of mobile Safari automatically submits the form when it
fills the credentials.

[1] https://lapcatsoftware.com/articles/Safari-autosubmit.html

In case of the Checkout page, it bypasses the redux-form provided submit
handler, ignoring the form's `onSubmit` prop, and does a non-SPA GET request to
`/check-out/delivery?<params>`, including the form values (username, password,
and "I agree to receive the newsletter") in the query parameters.  Which we
then post to Google Analytics.

From the data we gathered [2] the problem started occuring on 2021-08-27, which
is the date where, among other things, the form tag was introduced [3].

[2] https://docs.google.com/spreadsheets/d/1IMExIKFQJvX5rxMNxfJ3FNjpSoppoJMI/edit#gid=938755233
[3] https://github.com/Gousto/gousto-webclient/pull/3813/files

The current commit removes the form tag; this degrades the accesibility a
bit (submit on pressing Enter when focused in the username or password input
fields) but improves security.