[TG-5718] remove alt="animation" from Loading image (#5003)
* [TG-5718] remove alt="animation" from Loading image

At some point Google preview for the Gousto search was showing a spurious
section titled "Animation" and with the welcome promo code's text as
description.  The only externally-visible "animation" string is the "alt" text
on Loading; removing it probably will makes things better and definitely won't
make them worse.  (The usual guideline "images must have the `alt` attribute"
is not applicable here because the screen reader users will have an equally
hard time accessing the Gousto website whether or not this loader has the alt
text.)

In the past some Saved Pages on Optimizely (e.g. "GST 019") used
alt="animation" to refer to this element.  However there are no current
experiments using such Saved Pages; and we are migrating to Optimizely Rollouts
in any case, making this irrelevant.

* [TG-5718] when no class name, don't output "null" in the class attr