[TG-5717] display error page when orderId is requested by non-authenticated user (#5028)
In a July 2021 Technical Audit
https://docs.google.com/document/d/1L_Nu_NyAcvn2A_XJ7qX5Yq1Vss7sk2ky1oUp7tVcGwg/edit#

it was discovered that a search engine displays a link of a form
`gousto.co.uk/menu/<digits>` as one of the preview sections on the search
results page.  When a non-logged-in user accesses such a page, an infinite
loader was shown.  As a hint to the search engines, since the `menu/:orderId`
format is valid only for the logged-in users, for other cases we now display an
ErrorPage instead - like other "page not found" instances.