[TG-6412] fix retrieving promo code information for new users
It used the data dependency on the `promoCode` value of the result of the
`usePricing` hook; but this is incorrect because when there are not enough
recipes to checkout, that result is empty.  On the other hand, the basket slice
contains the promo code to apply even if there are no recipes; so we
cross-check the basket's promoCode with promoStore entry for that promo code.

It also is the same selector that's used for checking whether we should
`promoGet()` - retrieve the promoStore entry for that code - in

gousto-webclient/src/apps/webclient/src/routes/Menu/components/BoxSummary/Banner/BoxSummaryBanner.js:40