[TG-6412] refactor utilHooks; fix tip for existing users & no promo code
- remove unused properties from the type (used only in Price, which now gets
them directly from the usePricing hook response)
- unify two types because they now had the same essential data
- rename the type to DiscountDescriptor for clarity
- useDiscountTip change to formatDiscountTip because a pure function is
preferable to a hook

And the impetus for the change: there was a bug where if you sign up without a
promo code then go to Choose recipes, you see the tip as "£0 off youu box".
This is due to an incorrect API response that got into the pricing hook
response; which claims "flatDiscountApplied=true" with the amount=0.  Added
defensive checks around the actual discount amounts, so the UI considers a
discount enabled only when the amount is a string representing a positive
number.